requirements.txt
Django
django-leaflet
psycopg2